+++ src/workoutLogToExercise/workoutLogToExercise.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { WorkoutLogToExercise } from '../domain/WorkoutLogToExercise.entity';
+import { WorkoutLogToExercise } from './domain/WorkoutLogToExercise.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([WorkoutLogToExercise])],
